use 256

+++ contracts/src/challengeV2/EdgeChallengeManager.sol
@@ -172,7 +172,7 @@ interface IEdgeChallengeManager {
     ///         This value is increasing whilst an edge is unrivaled, once a rival is created
     ///         it is fixed. If an edge has rivals from the moment it is created then it will have
     ///         a zero time unrivaled
-    function timeUnrivaled(bytes32 edgeId) external view returns (uint64);
+    function timeUnrivaled(bytes32 edgeId) external view returns (uint256);
 
     /// @notice Get the id of the prev assertion that this edge is originates from
     /// @dev    Uses the parent chain to traverse upwards SmallStep->BigStep->Block->Assertion
@@ -236,7 +236,7 @@ contract EdgeChallengeManager is IEdgeChallengeManager, Initializable {
     /// @param edgeId               The edge that was confirmed
     /// @param mutualId             The mutual id of the confirmed edge
     /// @param totalTimeUnrivaled   The cumulative amount of time (in blocks) this edge spent unrivaled
-    event EdgeConfirmedByTime(bytes32 indexed edgeId, bytes32 indexed mutualId, uint64 totalTimeUnrivaled);
+    event EdgeConfirmedByTime(bytes32 indexed edgeId, bytes32 indexed mutualId, uint256 totalTimeUnrivaled);
 
     /// @notice An edge can be confirmed if a zero layer edge in the level below claims this edge
     /// @param edgeId           The edge that was confirmed
@@ -523,7 +523,7 @@ contract EdgeChallengeManager is IEdgeChallengeManager, Initializable {
             assertionBlocks = 0;
         }
 
-        uint64 totalTimeUnrivaled =
+        uint256 totalTimeUnrivaled =
             store.confirmEdgeByTime(edgeId, _unused, assertionBlocks, challengePeriodBlocks, NUM_BIGSTEP_LEVEL);
 
         emit EdgeConfirmedByTime(edgeId, store.edges[edgeId].mutualId(), totalTimeUnrivaled);
@@ -649,7 +649,7 @@ contract EdgeChallengeManager is IEdgeChallengeManager, Initializable {
     }
 
     /// @inheritdoc IEdgeChallengeManager
-    function timeUnrivaled(bytes32 edgeId) public view returns (uint64) {
+    function timeUnrivaled(bytes32 edgeId) public view returns (uint256) {
         return store.timeUnrivaled(edgeId);
     }
